requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
great_expectations==0.17.21
prometheus_client
pytest
flake8
pandera
apache-airflow==2.9.1          # Optional: only if you run tasks locally
dbt-core==1.7.9
dbt-postgres==1.7.9
